Re render

+++ src/Components/Workspace/Children/Task.jsx
@@ -4,8 +4,11 @@ import Tags from './Tags'
 import TextArea from './TextArea'
 import SpeechRecognition, { useSpeechRecognition } from 'react-speech-recognition'
 import moment from 'moment'
+import { confirmAlert } from 'react-confirm-alert'
+//import styles from './ConfirmStyles.module.css'
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
-const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTasksCount, decrementTasksCount }) => {
+const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, setOpen, getTasks }) => {
   const url = 'https://garage-best-team-ever.tk'
 
   const [on, setOn] = useState(true)
@@ -48,7 +51,7 @@ const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTas
 
   // Распознанная речь рендерится только в том таске, для которого включена запись
   const renderTranscript = () => {
-    if (!on) { return transcript }
+    if (!on)  {console.log(transcript); return transcript }
   }
 
   if (!SpeechRecognition.browserSupportsSpeechRecognition()) {
@@ -98,17 +101,23 @@ const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTas
 
   const saveTask = () => {
     const api = '/task'
-    const title = getTitle()
+    const title = getTitle().trim()
+    const textContent = text.trim()
+
+    if (title === '' && textContent == '')
+      return
+
     const tags = mapTags(allTags)
     let dateTarget = time
     if (typeof time === 'undefined' || time === '') {
       setTime(moment().add(1, 'days').startOf('hour').format('YYYY-MM-DD HH:mm:ss'))
       dateTarget = moment().add(1, 'days').startOf('hour').format('YYYY-MM-DD HH:mm:ss')
     }
+
     const data = {
       user_id: 0,
       title: title,
-      text_content: text.trim(),
+      text_content: textContent,
       date_target: dateTarget,
       tags: tags.map(tag => { return tag.name })
     }
@@ -122,7 +131,8 @@ const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTas
     })
       .then(() => setEditMode(false))
       .then(() => setIsNewTask(false))
-      .then(() => incrementTasksCount())
+      .then(() => getTasks())
+      .then(() => setOpen(false))
   }
 
   const updateTask = () => {
@@ -151,21 +161,44 @@ const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTas
       body: JSON.stringify(data)
     })
       .then(() => setEditMode(false))
+      .then(() => getTasks())
   }
 
   const deleteTask = () => {
-    const api = `/task/${taskId}`
+    if (taskId === undefined) {
+      setVisible(false)
+      setOpen(false)
+    } else {
+      const api = `/task/${taskId}`
 
-    fetch(url + api, {
-      method: 'DELETE'
-    })
-      .then(() => setVisible(false))
-      .then(() => decrementTasksCount())
+      fetch(url + api, {
+        method: 'DELETE'
+      })
+        .then(() => setVisible(false))
+        .then(() => getTasks())
+    }
   }
 
   //------------------------------------
   // Прочие функции
   //------------------------------------
+
+  // поп ап при удалении задачи
+  const deleteWarning = () => {
+    confirmAlert({
+      title: 'Подтверждение',
+      message: 'Вы уверены, что желаете удалить эту задачу?',
+      buttons: [
+        {
+          label: 'Да',
+          onClick: () => deleteTask()
+        },
+        {
+          label: 'Нет, я передумал :з',
+        }
+      ]
+    });
+  };
 
   // маппинг тегов в массив объектов
   const mapTags = tags => {
@@ -266,7 +299,7 @@ const Task = ({ id, title, bodyTask, tags, dateTarget, isNew, open, incrementTas
               <path d="M5,21h14c1.104,0,2-0.896,2-2V8l-5-5H5C3.896,3,3,3.896,3,5v14C3,20.104,3.896,21,5,21z M7,5h4v2h2V5h2v4h-1h-1h-2H9H7V5z M7,13h10v6h-2H9H7V13z" fill="#747E8A" />
             </svg>
           </button>
-          <button className={styles.DelIconContainer} onClick={deleteTask}>
+          <button className={styles.DelIconContainer} onClick={deleteWarning}>
             <svg className={styles.Icon + ' ' + styles.IconBottom} xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
               <path d="M5 8v12c0 1.104.896 2 2 2h10c1.104 0 2-.896 2-2V8c0 0-.447 0-1 0H6C5.447 8 5 8 5 8zM3 6L8 6 16 6 21 6 21 4 16.618 4 15 2 9 2 7.382 4 3 4z" fill="#747E8A" />
             </svg>
